requirements.txt
annotated-types==0.7.0
bandwidth-sdk==19.1.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pyngrok==7.2.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
